Online Course/requirements.txt
Flask
openai
python-dotenv
scikit-learn
pandas
numpy
pickle